Add  cheeseburger and coffee recipes
cheeseburger.html and coffee.html added to the file. Very simple
structuring lining up exactly with the existing lasagna.html
format. Several corresponding images got added to the images
directory as well. This includes a new homepage picture.

Beyond that, index.html had a minor change to the presention and
semantic html of the links. Nothing special, but it'll pay
dividends later.